♻️ Refactor : 내 픽셀 MOCK_DATA 구조 수정
